+++ cpsycle/ui/src/uilabel.c
@@ -75,7 +75,7 @@ void psy_ui_label_onpreferredsize(psy_ui_Label* self, psy_ui_Size* limit, psy_ui
 		} else {
 			rv->width = psy_ui_value_makepx(tm.tmAveCharWidth * self->charnumber);
 		}
-		rv->height = psy_ui_value_makepx((tm.tmHeight * self->linespacing) +
+		rv->height = psy_ui_value_makepx((int)(tm.tmHeight * self->linespacing) +
 			psy_ui_margin_height_px(&psy_ui_label_base(self)->spacing, &tm));
 	}
 }
